Abstract out interactions plot and add tooltip
